+++ packages/sdk/src/transactions/Payload.ts
@@ -611,10 +611,10 @@ export function deserialize(value: Cursor | ArrayBuffer): Payload {
             payload = registerData(getAccountTransactionHandler(type).deserialize(cursor));
             break;
         case AccountTransactionType.ConfigureDelegation:
-            payload = configureDelegation(getAccountTransactionHandler(type).deserialize());
+            payload = configureDelegation(getAccountTransactionHandler(type).deserialize(cursor));
             break;
         case AccountTransactionType.ConfigureBaker:
-            payload = configureValidator(getAccountTransactionHandler(type).deserialize());
+            payload = configureValidator(getAccountTransactionHandler(type).deserialize(cursor));
             break;
         case AccountTransactionType.TokenUpdate:
             payload = tokenUpdate(getAccountTransactionHandler(type).deserialize(cursor));